styled and added userApp

+++ src/blogapp/views.py
@@ -60,6 +60,10 @@ def post_update(request, slug):
 
 def post_delete(request, slug):
     obj = get_object_or_404(Post, slug=slug)
+    if request.user.id != obj.author.id :
+        return HttpResponse("You're not authorized!!")
+
+
     if request.method =="POST":
         obj.delete()
         return redirect("blogapp:list")
